+++ demo/views/forms/forms.js
@@ -2,7 +2,11 @@ import React from 'react';
 import TextboxDemo from './textbox-demo';
 import DecimalDemo from './decimal-demo';
 import DropdownDemo from './dropdown-demo';
+import DropdownFilterDemo from './dropdown-filter-demo';
+import DropdownFilterAjaxDemo from './dropdown-filter-ajax-demo';
 import DateDemo from './date-demo';
+import CheckboxDemo from './checkbox-demo';
+import RadioButtonDemo from './radio-button-demo';
 
 class Forms extends React.Component {
   /**
@@ -12,10 +16,13 @@ class Forms extends React.Component {
     return (
       <div>
         <h1>Forms</h1>
+        <CheckboxDemo />
+        <RadioButtonDemo/>
+        <TextboxDemo />
         <DecimalDemo />
         <DateDemo />
         <DropdownDemo />
-        <TextboxDemo />
+        <DropdownFilterDemo />
       </div>
     );
   }