+++ demo/views/modals/modals.js
@@ -1,6 +1,9 @@
 import React from 'react';
 import DialogDemo from './dialog-demo';
 import DialogFullScreenDemo from './dialog-full-screen-demo';
+import SidebarDemo from './sidebar-demo';
+import AlertDemo from './alert-demo';
+import ConfirmDemo from './confirm-demo';
 
 class Modals extends React.Component {
   /**
@@ -12,6 +15,9 @@ class Modals extends React.Component {
         <h1>Modals</h1>
         <DialogDemo />
         <DialogFullScreenDemo />
+        <SidebarDemo />
+        <AlertDemo />
+        <ConfirmDemo />
       </div>
     );
   }