+++ src/components/radio-button/radio-button.js
@@ -131,6 +131,7 @@ class RadioButton extends React.Component {
       <div className={ this.mainClasses }>
         { this.inputHTML }
         { this.labelHTML }
+        { this.labelHelpHTML }
         { this.validationHTML }
       </div>
     );